requirements.txt
pytest~=7.4.0
selene>=2.0.0rc4
allure-pytest == 2.13.2
python-dotenv>=0.21.0
selenium~=4.10.0
pydantic==2.4.2
pydantic-settings==2.0.3